moved data fetching to top level of profile page --- added loading spinner focus window to profile, incidents, and incident details pages
