+++ app/_components/Table.tsx
@@ -11,6 +11,8 @@ import {
 import { useSession } from 'next-auth/react';
 import { useRouter } from 'next/navigation';
 import { Incident, UserName } from '../../types/definitions';
+import Backdrop from '@mui/material/Backdrop';
+import CircularProgress from '@mui/material/CircularProgress';
 
 
 
@@ -133,8 +135,16 @@ const Table = () => {
   ];
 
   while (loading) {
-    return <div>loading incidents ...</div>
-  }
+    return ( 
+    <div>
+      <Backdrop
+        sx={{ color: '#fff', zIndex: (theme) => theme.zIndex.drawer + 1 }}
+        open={true}
+      >
+        <CircularProgress color="inherit" />
+      </Backdrop>
+    </div>
+  )}
 
 
   return (
