+++ app/_components/userPreferences/DeleteAccount.tsx
@@ -1,22 +1,18 @@
 'use client'
 
 import * as React from 'react';
-import { useSession } from 'next-auth/react';
 import { useRouter } from 'next/navigation';
 import { signOut } from "next-auth/react";
 
-const DeleteAccount = () => {
+const DeleteAccount = (props: {user_id: (string | undefined)}) => {
 
     const router = useRouter();
-
-    const session = useSession().data;
-    const userId = session?.user?.userid;
 
     function handleClick() {
         if (confirm('Deleting user account will remove all data associated with this account. This action cannot be undone. Do you wish to continue deleting your account?')) {
             fetch('http://localhost:3000/api/updateUser/removeAccount', {
                 method: 'POST',
-                body: JSON.stringify({user_id: userId})
+                body: JSON.stringify({user_id: props.user_id})
             })
             signOut();
         }
